Factor out auto-translation code
