+++ app/models/place.rb
@@ -1,12 +1,14 @@
-require 'auto_translator'
+require 'place/auto_translate'
 
 class Place < ActiveRecord::Base
+  include PlaceAutoTranslation
+
   def self.reviewed
     Place.all.map(&:reviewed_version).compact
   end
 
   def self.with_reviewed_category(id)
-    Place.all.map(&:reviewed_version).compact.find_all{ |p| p.has_category?(id) }
+    Place.all.map(&:reviewed_version).compact.find_all { |p| p.has_category?(id) }
   end
 
   ## VALIDATIONS
@@ -14,7 +16,7 @@ class Place < ActiveRecord::Base
   validates :postal_code, format: { with: /\d{5}/, message: 'supply valid postal code (5 digits)' }
 
   ## TRANSLATION
-  translates :description, versioning: { gem: :paper_trail, options: { on: [ :update, :create ] } }
+  translates :description, versioning: { gem: :paper_trail, options: { on: [:update, :create] } }
   globalize_accessors
 
   ## CALLBACKS
@@ -37,53 +39,6 @@ class Place < ActiveRecord::Base
 
   def unreviewed_version
     self if versions.length > 1 || (versions.length == 1 && !reviewed)
-  end
-
-  ## Language and autotranslation related stuff
-  # Maybe refactor
-  # Why obj.split(' ').length == 1 ?? raises error in case of one-word-description (no error handling)
-  def emptyish?(obj)
-    obj.nil? || obj.empty? || obj.split(' ').length == 1
-  end
-
-  def autotranslated_or_empty_descriptions
-    translations.select { |t| t.auto_translated || emptyish?(t.description) }
-  end
-
-  def locales_of_empty_descriptions
-    autotranslated_or_empty_descriptions.map(&:locale)
-  end
-
-  def translations_with_descriptions
-    translations - autotranslated_or_empty_descriptions
-  end
-
-  def guess_native_language_description
-    translations_with_descriptions.sort_by do |t|
-      t.description.length
-    end.last
-  end
-
-  def translate_empty_descriptions
-    locales_of_empty_descriptions.each do |missing_locale|
-      auto_translation = @translator.failsafe_translate(
-        @native_translation.description,
-        @native_translation.locale.to_s,
-        missing_locale.to_s
-      )
-      translation = translations.find_by(locale: missing_locale)
-      translation.without_versioning do
-        translation.update_attributes(description: auto_translation,
-                                      auto_translated: true,
-                                      reviewed: false)
-      end
-    end
-  end
-
-  def auto_translate
-    @native_translation = guess_native_language_description
-    @translator = BingTranslatorWrapper.new(ENV['bing_id'], ENV['bing_secret'], ENV['microsoft_account_key'])
-    translate_empty_descriptions if @translator && @native_translation
   end
 
   ## CATEGORIES
